Разделила колонки и карточки
